+++ src/main/java/usyd/mingyi/animalcare/mapper/FriendMapper.java
@@ -14,7 +14,7 @@ public interface FriendMapper {
     * @Author: Mingyi Li
     * @Date: 2022/10/5 
     */ 
-    int checkExistRequest(int fromId,int toId);
+    int checkExistRequest(long fromId,long toId);
     /** 
     * @Description:  检查是否被该用户请求添加过
     * @Param: [fromId, toId] 
@@ -22,7 +22,7 @@ public interface FriendMapper {
     * @Author: Mingyi Li
     * @Date: 2022/10/5 
     */ 
-    int checkRequestReverse(int fromId,int toId);
+    int checkRequestReverse(long fromId,long toId);
     /** 
     * @Description:  发送到对方请求列表中
     * @Param: [fromId, toId] 
@@ -30,7 +30,7 @@ public interface FriendMapper {
     * @Author: Mingyi Li
     * @Date: 2022/10/5 
     */ 
-    int sendFriendRequest(int fromId,int toId);
+    int sendFriendRequest(long fromId,long toId);
     /** 
     * @Description:  在请求列表中删除请求
     * @Param: [fromId, toId] 
@@ -38,7 +38,7 @@ public interface FriendMapper {
     * @Author: Mingyi Li
     * @Date: 2022/10/5 
     */ 
-    int deleteFriendRequest(int fromId,int toId);//
+    int deleteFriendRequest(long fromId,long toId);//
     
     /** 
     * @Description:  互相添加到好友列表
@@ -47,7 +47,7 @@ public interface FriendMapper {
     * @Author: Mingyi Li
     * @Date: 2022/10/5 
     */ 
-    int addToFriendList(int fromId,int toId);
+    int addToFriendList(long fromId,long toId);
     
     /** 
     * @Description: 判断是否已经是好友了
@@ -57,7 +57,7 @@ public interface FriendMapper {
     * @Date: 2022/10/5 
     */
     
-    boolean isFriend(int fromId,int toId);
+    boolean isFriend(long fromId,long toId);
     /** 
     * @Description:  根据用户id获取用户所有的friends
     * @Param: [id] 
@@ -65,7 +65,7 @@ public interface FriendMapper {
     * @Author: Mingyi Li
     * @Date: 2022/10/5 
     */ 
-    List<User> getAllFriends(int id);
+    List<User> getAllFriends(long id);
     /** 
     * @Description: 根据用户id获取用户所有的requests
     * @Param: [id] 
@@ -74,7 +74,7 @@ public interface FriendMapper {
     * @Date: 2022/10/5 
     */
 
-    List<User> getAllRequests(int id);
+    List<User> getAllRequests(long id);
 
 
     /** 
@@ -84,7 +84,7 @@ public interface FriendMapper {
     * @Author: Mingyi Li
     * @Date: 2022/10/8 
     */ 
-    int deleteFromFriendList(int fromId,int toId);
+    int deleteFromFriendList(long fromId,long toId);
     
 
 }
